+++ services/kaggle-data-api/src/main.ts
@@ -1,23 +1,19 @@
-import Fastify from 'fastify';
-import { app } from './app/app.js';
+import { build } from './app/app.js';
 
 const host = process.env.HOST ?? 'localhost';
 const port = process.env.PORT ? Number(process.env.PORT) : 3000;
 
-// Instantiate Fastify with some config
-const server = Fastify({
-  logger: true,
-});
+async function start() {
+  const server = await build({});
+  server.log.level = process.env.LOG_LEVEL ?? 'info';
+  server.listen({ port, host }, (err) => {
+    if (err) {
+      server.log.error(err);
+      process.exit(1);
+    } else {
+      console.log(`[ ready ] http://${host}:${port}`);
+    }
+  });
+}
 
-// Register your application as a normal plugin.
-server.register(app);
-
-// Start listening.
-server.listen({ port, host }, (err) => {
-  if (err) {
-    server.log.error(err);
-    process.exit(1);
-  } else {
-    console.log(`[ ready ] http://${host}:${port}`);
-  }
-});
+start();